clean up, renamed files for more clarity
